RBFModels

RBFModels.RBFModel

RBFModels.RBFOutputSystem

RBFModels.RBFInterpolationModel

RBFModels._distances

RBFModels._eval_rbfs_at_ρ

RBFModels._func_matrix

RBFModels._kernel_vector

RBFModels.canonical_basis

RBFModels.coefficients

RBFModels.make_kernels

RBFModels.make_kernels

RBFModels.non_negative_solutions

RBFModels.RBFModel — Type

RBFModel{S,V}

S is true or false and indicates whether static arrays are used or not.

V is true if vectors should be returned and false if scalars are returned.

Initialize via one of the constructors, e.g., RBFInterpolationModel( sites, values, φ, poly_deg ) to obain an interpolating RBF model.

See also RBFInterpolationModel

RBFModels.RBFOutputSystem — Type

Evaluate rbf :: RBFOutputSystem at site x.

RBFModels.RBFInterpolationModel — Method

RBFInterpolationModel( sites :: Vector{VS}, values :: Vector{VT}, φ, poly_deg = 1; 
    static_arrays = nothing, vector_output = true ) where {VS<:NumberOrVector, VT<:NumberOrVector}

Return an RBFModel m that is interpolating, i.e., m(sites[i]) == values[i] for all i = eachindex(sites). φ should be a RadialFunction or a vector of RadialFunctions that has the same length as sites and values. poly_deg specifies the degree of the multivariate polynomial added to the RBF model. It will be reset if needed. static_arrays is automatically set to true if unspecified and the data dimensions are small. vector_output is ignored if the values have length > 1. Elsewise it specifies whether to return vectors or scalars when evaluating.

RBFModels._distances — Method

Return a vector containing the distance of x to each kernel center of RBFOutputSystem.

RBFModels._eval_rbfs_at_ρ — Function

Evaluate (output ℓ of) rbf by plugging in distance ρ[i] in radial function o.kernels[i].φ.

RBFModels._func_matrix — Method

Evaluate each function in funcs on each number/vector in func_args, so that each column corresponds to a function evaluation.

RBFModels._kernel_vector — Method

Return the vector [ φ₁(ρ₁) … φₙ(ρₙ) ] = [ k1(x) … kn(x) ]

RBFModels.canonical_basis — Method

canonical_basis( n:: Int, d :: Int )

Return the canonical basis of the space of n-variate polynomials of degree at most d.

RBFModels.coefficients — Method

coefficients(sites, values, centers, rad_funcs, polys )

Return the coefficient matrices w and λ for an rbf model $r(x) = Σ_{i=1}^N wᵢ φ(\|x - x^i\|) + Σ_{j=1}^M λᵢ pᵢ(x)$, where $N$ is the length of rad_funcs (and centers) and $M$ is the length of polys.

The arguments are

an array of data sites sites with vector entries from $ℝ^n$.

an array of data values values with vector entries from $ℝ^k$.

an array of ShiftedKernels.

an array polys of polynomial basis functions.

RBFModels.make_kernels — Method

Return array of ShiftedKernels based on φ with centers from sites.

RBFModels.make_kernels — Method

Return array of ShiftedKernels based functions in φ_arr with centers from sites.

RBFModels.non_negative_solutions — Method

non_negative_solutions( d :: Int, n :: Int)

Return array of solution vectors [x1, …, xn] to the equation $x_1 + … + x_n = d$ where the variables are non-negative integers.