RadialBasisFunctionModels

Base.AbstractVector

RadialBasisFunctionModels.Cubic

RadialBasisFunctionModels.EmptyPolySystem

RadialBasisFunctionModels.Gaussian

RadialBasisFunctionModels.InverseMultiquadric

RadialBasisFunctionModels.Multiquadric

RadialBasisFunctionModels.RBFInterpolationModel

RadialBasisFunctionModels.RBFModel

RadialBasisFunctionModels.RBFModel

RadialBasisFunctionModels.RBFModel

RadialBasisFunctionModels.RBFModel

RadialBasisFunctionModels.RBFSum

RadialBasisFunctionModels.RBFSum

RadialBasisFunctionModels.ShiftedKernel

RadialBasisFunctionModels.ThinPlateSpline

RadialBasisFunctionModels.auto_grad

RadialBasisFunctionModels.auto_jac

RadialBasisFunctionModels.coefficients

RadialBasisFunctionModels.eval_and_auto_grad

RadialBasisFunctionModels.eval_and_auto_jac

RadialBasisFunctionModels.eval_at_dist

RadialBasisFunctionModels.eval_at_dist

RadialBasisFunctionModels.make_kernels

RadialBasisFunctionModels.make_kernels

RadialBasisFunctionModels.non_negative_solutions

StaticPolynomials.evaluate

StaticPolynomials.evaluate

ZygoteRules._pullback

Base.AbstractVector — Method

Evaluate $x ↦ [ k₁(x), …, k_{N_c}(x)]$ at x.

RadialBasisFunctionModels.Cubic — Type

Cubic( β = 3 ) <: RadialFunction

A RadialFunction with

\[ φ(ρ) = (-1)^{ \lceil β \rceil /2 } ρ^β\]

RadialBasisFunctionModels.EmptyPolySystem — Type

Drop-In Alternative to StaticPolynomials.PolynomialSystem when there are no outputs.

RadialBasisFunctionModels.Gaussian — Type

Gaussian( α = 1 ) <: RadialFunction

A RadialFunction with

\[ φ(ρ) = \exp( - (α ρ)^2 ).\]

RadialBasisFunctionModels.InverseMultiquadric — Type

InverseMultiquadric( α = 1, β = 1//2 ) <: RadialFunction

A RadialFunction with

\[ φ(ρ) = ( 1 + (αρ)^2 )^{-β}\]

RadialBasisFunctionModels.Multiquadric — Type

Multiquadric( α = 1, β = 1//2 ) <: RadialFunction

A RadialFunction with

\[ φ(ρ) = (-1)^{ \lceil β \rceil } ( 1 + (αρ)^2 )^β\]

RadialBasisFunctionModels.RBFInterpolationModel — Type

RBFInterpolationModel(features, labels, φ, poly_deg; kwargs… )

Build a model interpolating the feature-label pairs. Does not accept center keyword argument.

RadialBasisFunctionModels.RBFModel — Type

RBFModel{V}

V is true by default. It can be set to false only if the number of outputs is 1. Then scalars are returned.

RadialBasisFunctionModels.RBFModel — Type

Evaluate model mod :: RBFModel at vector x.

RadialBasisFunctionModels.RBFModel — Type

RBFModel( features, labels, φ = Multiquadric(), poly_deg = 1; kwargs ... )

Construct a RBFModel from the feature vectors in features and the corresponding labels in lables, where φ is a RadialFunction or a vector of RadialFunctions.

Scalar data can be used, it is transformed internally.

StaticArrays can be used, e.g., features :: Vector{<:SVector}. Providing SVectors only might speed up the construction.

If the degree of the polynomial tail, poly_deg, is too small it will be set to cpd_order(φ)-1.

If the RBF centers do not equal the the features, you can use the keyword argument centers to pass a list of centers. If φ is a vector, then the length of centers and φ must be equal and centers[i] will be used in conjunction with φ[i] to build a ShiftedKernel.

If features has 1D data, the output of the model will be a 1D-vector. If it should be a scalar instead, set the keyword argument vector_output to false.

RadialBasisFunctionModels.RBFModel — Method

Evaluate scalar output ℓ of model mod at vector x.

RadialBasisFunctionModels.RBFSum — Method

Evaluate output ℓ of RBF sum rbf::RBFSum

RadialBasisFunctionModels.RBFSum — Method

Evaluate rbf::RBFSum at x.

RadialBasisFunctionModels.ShiftedKernel — Method

Evaluate kernel k at x - k.c.

RadialBasisFunctionModels.ThinPlateSpline — Type

ThinPlateSpline( k = 2 ) <: RadialFunction

A RadialFunction with

\[ φ(ρ) = (-1)^{k+1} ρ^{2k} \log(ρ)\]

RadialBasisFunctionModels.auto_grad — Function

Return gradient of output ℓ of model rbf at point x (using Zygote).

RadialBasisFunctionModels.auto_jac — Method

Return the jacobian of rbf at x (using Zygote).

RadialBasisFunctionModels.coefficients — Method

coefficients(sites, values, kernels, rad_funcs, polys )

Return the coefficient matrices w and λ for an rbf model $r(x) = Σ_{i=1}^N wᵢ φ(\|x - x^i\|) + Σ_{j=1}^M λᵢ pᵢ(x)$, where $N$ is the length of rad_funcs (and centers) and $M$ is the length of polys.

The arguments are

an array of data sites sites with vector entries from $ℝ^n$.

an array of data values values with vector entries from $ℝ^k$.

an array of ShiftedKernels.

a PolynomialSystem or EmptyPolySystem (in case of deg = -1).

RadialBasisFunctionModels.eval_and_auto_grad — Function

Evaluate output ℓ of the model and return the gradient.

RadialBasisFunctionModels.eval_and_auto_jac — Method

Evaluate the model and return the jacobian at the same time.

RadialBasisFunctionModels.eval_at_dist — Method

Evaluate $x ↦ [ k₁(x), …, k_{N_c}(x)]$, provided the distances $[ ρ_1(x), …, ρ_{N_c}(x) ]$.

RadialBasisFunctionModels.eval_at_dist — Method

Evaluate k.φ for distance ρ where ρ should equal x - k.c for the argument x.

RadialBasisFunctionModels.make_kernels — Method

Return array of ShiftedKernels based functions in φ_arr with centers from centers.

RadialBasisFunctionModels.make_kernels — Method

Return array of ShiftedKernels based function φ with centers from centers.

RadialBasisFunctionModels.non_negative_solutions — Method

non_negative_solutions( d :: Int, n :: Int)

Return a matrix with columns that correspond to solution vectors $[x_1, …, x_n]$ to the equation $x_1 + … + x_n = d$, where the variables are non-negative integers.

StaticPolynomials.evaluate — Method

Evaluate for sized input.

StaticPolynomials.evaluate — Method

Evaluate for usual vector input. (Scalar input also supported, there are no checks)

ZygoteRules._pullback — Method

canonical_basis( n:: Int, d :: Int ) :: Union{PolynomialSystem, EmptyPolySystem}

Return the canonical basis of the space of n-variate polynomials of degree at most d.