RBFModels

RBFModels.AnyVec

RBFModels.Cubic

RBFModels.Gaussian

RBFModels.InverseMultiquadric

RBFModels.Multiquadric

RBFModels.RBFInterpolationModel

RBFModels.RBFModel

RBFModels.RBFModel

RBFModels.RBFModel

RBFModels.RBFModel

RBFModels.RBFSum

RBFModels.RBFSum

RBFModels.ShiftedKernel

RBFModels.ThinPlateSpline

RBFModels.auto_grad

RBFModels.auto_jac

RBFModels.canonical_basis

RBFModels.coefficients

RBFModels.eval_and_auto_grad

RBFModels.eval_and_auto_jac

RBFModels.eval_at_dist

RBFModels.eval_at_dist

RBFModels.make_kernels

RBFModels.make_kernels

RBFModels.non_negative_solutions

RBFModels.AnyVec — Method

Evaluate $x ↦ [ k₁(x), …, k_{N_c}(x)]$ at x.

RBFModels.Cubic — Type

Cubic( β = 3 ) <: RadialFunction

A RadialFunction with

\[ φ(ρ) = (-1)^{ \lceil β \rceil /2 } ρ^β\]

RBFModels.Gaussian — Type

Gaussian( α = 1 ) <: RadialFunction

A RadialFunction with

\[ φ(ρ) = \exp( - (α ρ)^2 ).\]

RBFModels.InverseMultiquadric — Type

InverseMultiquadric( α = 1, β = 1//2 ) <: RadialFunction

A RadialFunction with

\[ φ(ρ) = ( 1 + (αρ)^2 )^{-β}\]

RBFModels.Multiquadric — Type

Multiquadric( α = 1, β = 1//2 ) <: RadialFunction

A RadialFunction with

\[ φ(ρ) = (-1)^{ \lceil β \rceil } ( 1 + (αρ)^2 )^β\]

RBFModels.RBFInterpolationModel — Type

RBFInterpolationModel(features, labels, φ, poly_deg; kwargs… )

Build a model interpolating the feature-label pairs. Does not accept center keyword argument.

RBFModels.RBFModel — Type

RBFModel{V}

V is true by default. It can be set to false only if the number of outputs is 1. Then scalars are returned.

RBFModels.RBFModel — Type

Evaluate model mod :: RBFModel at vector x.

RBFModels.RBFModel — Type

RBFModel( features, labels, φ = Multiquadric(), poly_deg = 1; kwargs ... )

Construct a RBFModel from the feature vectors in features and the corresponding labels in lables, where φ is a RadialFunction or a vector of RadialFunctions.

Scalar data can be used, it is transformed internally.

StaticArrays can be used, e.g., features :: Vector{<:SVector}. Providing SVectors only might speed up the construction.

If the degree of the polynomial tail, poly_deg, is too small it will be set to cpd_order(φ)-1.

If the RBF centers do not equal the the features, you can use the keyword argument centers to pass a list of centers. If φ is a vector, then the length of centers and φ must be equal and centers[i] will be used in conjunction with φ[i] to build a ShiftedKernel.

If features has 1D data, the output of the model will be a 1D-vector. If it should be a scalar instead, set the keyword argument vector_output to false.

RBFModels.RBFModel — Method

Evaluate scalar output ℓ of model mod at vector x.

RBFModels.RBFSum — Method

Evaluate output ℓ of RBF sum rbf::RBFSum

RBFModels.RBFSum — Method

Evaluate rbf::RBFSum at x.

RBFModels.ShiftedKernel — Method

Evaluate kernel k at x - k.c.

RBFModels.ThinPlateSpline — Type

ThinPlateSpline( k = 2 ) <: RadialFunction

A RadialFunction with

\[ φ(ρ) = (-1)^{k+1} ρ^{2k} \log(ρ)\]

RBFModels.auto_grad — Function

Return gradient of output ℓ of model rbf at point x (using Zygote).

RBFModels.auto_jac — Method

Return the jacobian of rbf at x (using Zygote).

RBFModels.canonical_basis — Method

canonical_basis( n:: Int, d :: Int )

Return the canonical basis of the space of n-variate polynomials of degree at most d.

RBFModels.coefficients — Method

coefficients(sites, values, kernels, rad_funcs, polys )

Return the coefficient matrices w and λ for an rbf model $r(x) = Σ_{i=1}^N wᵢ φ(\|x - x^i\|) + Σ_{j=1}^M λᵢ pᵢ(x)$, where $N$ is the length of rad_funcs (and centers) and $M$ is the length of polys.

The arguments are

an array of data sites sites with vector entries from $ℝ^n$.

an array of data values values with vector entries from $ℝ^k$.

an array of ShiftedKernels.

an array polys of polynomial basis functions.

RBFModels.eval_and_auto_grad — Function

Evaluate output ℓ of the model and return the gradient.

RBFModels.eval_and_auto_jac — Method

Evaluate the model and return the jacobian at the same time.

RBFModels.eval_at_dist — Method

Evaluate k.φ for distance ρ where ρ should equal x - k.c for the argument x.

RBFModels.eval_at_dist — Method

Evaluate $x ↦ [ k₁(x), …, k_{N_c}(x)]$, provided the distances $[ ρ_1(x), …, ρ_{N_c}(x) ]$.

RBFModels.make_kernels — Method

Return array of ShiftedKernels based function φ with centers from centers.

RBFModels.make_kernels — Method

Return array of ShiftedKernels based functions in φ_arr with centers from centers.

RBFModels.non_negative_solutions — Method

non_negative_solutions( d :: Int, n :: Int)

Return array of solution vectors $[x_1, …, x_n]$ to the equation $x_1 + … + x_n = d$ where the variables are non-negative integers.