RBFModels

RBFModels.RBFInterpolationModel

RBFModels.canonical_basis

RBFModels.coefficients

RBFModels.make_kernels

RBFModels.make_kernels

RBFModels.non_negative_solutions

RBFModels.RBFInterpolationModel — Method

Evaluate m at vector x, using size information.

RBFModels.canonical_basis — Method

canonical_basis( n:: Int, d :: Int )

Return the canonical basis of the space of n-variate polynomials of degree at most d.

RBFModels.coefficients — Method

coefficients(sites, values, kernels, polys )

Return the coefficient matrices w and λ for an rbf model $r(x) = Σ_{i=1}^N wᵢ kᵢ(x) + Σ_{j=1}^M λᵢ pᵢ(x)$, where $N$ is the length of kernels and $M$ is the length of polys.

The arguments are

an array of data sites sites with vector entries from $ℝ^n$.

an array of data values values with vector entries from $ℝ^k$.

an array kernels of RBF basis functions (of type ShiftedKernel)

an array polys of polynomial basis functions.

RBFModels.make_kernels — Method

Return array of ShiftedKernels based on φ with centers from sites.

RBFModels.make_kernels — Method

Return array of ShiftedKernels based functions in φ_arr with centers from sites.

RBFModels.non_negative_solutions — Method

non_negative_solutions( d :: Int, n :: Int)

Return array of solution vectors [x1, …, xn] to the equation $x_1 + … + x_n = d$ where the variables are non-negative integers.